refactor: Move integration test setup to `data_hub_shared`
